fixed comparing int and unsinged long
changed int to size_t

+++ src/board.h
@@ -120,7 +120,7 @@ public:
     int get_from_csv(std::string line,int n){
         int x = 0;
         int j = 0;
-        for(int i = 0;i < line.size();i++){
+        for(size_t i = 0;i < line.size();i++){
             if(line[i] == ';'){
                 j++;
                 continue;
